+++ tests/integration/model_providers/openai/test_openai.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import asyncio
 import os
 import time
 import unittest.mock
@@ -32,9 +31,12 @@ from mlrun.datastore.datastore_profile import (
     OpenAIProfile,
     register_temporary_client_datastore_profile,
 )
-from mlrun.datastore.model_provider.model_provider import ModelProvider
 from mlrun.datastore.model_provider.openai_provider import OpenAIProvider
-from tests.datastore.remote_model.remote_model_utils import setup_remote_model_test, INPUT_DATA, EXPECTED_RESULTS
+from mlrun.serving import ModelRunnerStep
+from tests.datastore.remote_model.remote_model_utils import (
+    EXPECTED_RESULTS,
+    INPUT_DATA,
+)
 
 here = os.path.dirname(__file__)
 config = {}
@@ -42,61 +44,6 @@ config_file_path = os.path.join(here, "test-openai.yml")
 if os.path.exists(config_file_path):
     with open(config_file_path) as yaml_file:
         config = yaml.safe_load(yaml_file).get("env", {})
-
-
-async def timed(coro):
-    start = time.perf_counter()
-    result = await coro
-    duration = time.perf_counter() - start
-    return result, duration
-
-
-class MyOpenAILLM(mlrun.serving.states.Model):
-    def predict(self, body):
-        if isinstance(
-            self.invocation_artifact, mlrun.artifacts.LLMPromptArtifact
-        ) and isinstance(self.model_provider, ModelProvider):
-            prompt = self.enrich_prompt(body)
-            body["result"] = self.model_provider.invoke(
-                prompt=prompt,
-                **(self.invocation_artifact.spec.model_configuration or {}),
-            )
-        return body
-
-    async def predict_async(self, body):
-        if isinstance(
-            self.invocation_artifact, mlrun.artifacts.LLMPromptArtifact
-        ) and isinstance(self.model_provider, ModelProvider):
-            prompt_parameters: list = body["input"]
-            prompts = [
-                self.enrich_prompt(single_prompt_parameters)
-                for single_prompt_parameters in prompt_parameters
-            ]
-
-            tasks = [
-                timed(
-                    self.model_provider.async_invoke(
-                        prompt,
-                        **(self.invocation_artifact.spec.model_configuration or {}),
-                    )
-                )
-                for prompt in prompts
-            ]
-            results_with_times = await asyncio.gather(*tasks)
-            results = [r for r, _ in results_with_times]
-            invoke_times = [t for _, t in results_with_times]
-            body["results"] = results
-            body["invoke_times"] = invoke_times
-        return body
-
-    def enrich_prompt(self, body) -> str:
-        # TODO: Update this once ML-8172 is completed
-        if isinstance(self.invocation_artifact, mlrun.artifacts.LLMPromptArtifact):
-            prompt_template = self.invocation_artifact.spec.prompt_string
-            needed_params = ["question", "depth_level", "persona", "tone"]
-            sub_dict = {k: body[k] for k in needed_params if k in body}
-            return prompt_template.format(**sub_dict)
-        return ""
 
 
 def create_mocked_get_store_artifact(uri_to_artifact: dict):
@@ -293,73 +240,30 @@ class TestOpenAIProvider(TestBasicOpenAIProvider):
 
 
 class TestOpenAIModel(TestBasicOpenAIProvider):
-    # @pytest.fixture
-    # def prompt_data(self):
-    #     return {
-    #         "input": [
-    #             {
-    #                 "question": "What is the capital of France, and give a brief historical overview.",
-    #                 "depth_level": "detailed",
-    #                 "persona": "teacher",
-    #                 "tone": "casual",
-    #             },
-    #             {
-    #                 "question": "What is 2 + 2? Answer shortly and then explain with details.",
-    #                 "depth_level": "basic",
-    #                 "persona": "math teacher",
-    #                 "tone": "simple",
-    #             },
-    #             {
-    #                 "question": "Who wrote Hamlet? Answer shortly and then explain with details.",
-    #                 "depth_level": "basic",
-    #                 "persona": "literature professor",
-    #                 "tone": "formal",
-    #             },
-    #             {
-    #                 "question": "What color is the sky on a clear day? Answer shortly and then explain with details.",
-    #                 "depth_level": "basic",
-    #                 "persona": "child",
-    #                 "tone": "fun",
-    #             },
-    #             {
-    #                 "question": "What planet do we live on? Answer shortly and then explain with details.",
-    #                 "depth_level": "basic",
-    #                 "persona": "astronaut",
-    #                 "tone": "educational",
-    #             },
-    #         ],
-    #     }
-    # 
-    # @pytest.fixture
-    # def prompt_expected_results(self):
-    #     return ["paris", "4", "shakespeare", "blue", "earth"]
-
-    # def setup_remote_model_test(self):
-    #     project = mlrun.new_project("test-openai-model", save=False)
-    #     model_url = self.url_prefix + self.basic_llm_model
-    #     model_artifact = project.log_model(
-    #         "my_model",
-    #         model_url=model_url,
-    #         default_config={"max_tokens": 100},
-    #     )
-    #     prompt_template = (
-    #         "{question}. Explain {depth_level} as a {persona} in {tone} style."
-    #     )
-    #     llm_prompt_artifact = project.log_llm_prompt(
-    #         "my_llm_prompt",
-    #         prompt_string=prompt_template,
-    #         model_artifact=model_artifact.uri,
-    #     )
-    #     function = mlrun.new_function("tests", kind="serving")
-    #     graph = function.set_topology("flow", engine="async")
-    #     model_runner_step = ModelRunnerStep(name="my_model_runner")
-    #     return model_artifact, llm_prompt_artifact, function, graph, model_runner_step
-
-    def test_model_runner_with_openai(self):
+    def _setup_remote_model_test(self):
         project = mlrun.new_project("test-openai-model", save=False)
         model_url = self.url_prefix + self.basic_llm_model
+        model_artifact = project.log_model(
+            "my_model",
+            model_url=model_url,
+            default_config={"max_tokens": 100},
+        )
+        prompt_template = (
+            "{question}. Explain {depth_level} as a {persona} in {tone} style."
+        )
+        llm_prompt_artifact = project.log_llm_prompt(
+            "my_llm_prompt",
+            prompt_string=prompt_template,
+            model_artifact=model_artifact.uri,
+        )
+        function = mlrun.new_function("tests", kind="serving")
+        graph = function.set_topology("flow", engine="async")
+        model_runner_step = ModelRunnerStep(name="my_model_runner")
+        return model_artifact, llm_prompt_artifact, function, graph, model_runner_step
+
+    def test_model_runner_with_openai(self):
         model_artifact, llm_prompt_artifact, function, graph, model_runner_step = (
-            setup_remote_model_test(project=project, model_url=model_url)
+            self._setup_remote_model_test()
         )
         model_runner_step.add_model(
             model_class="MyOpenAILLM",
@@ -395,10 +299,8 @@ class TestOpenAIModel(TestBasicOpenAIProvider):
             server.wait_for_completion()
 
     def test_model_runner_with_openai_async(self):
-        project = mlrun.new_project("test-openai-model", save=False)
-        model_url = self.url_prefix + self.basic_llm_model
         model_artifact, llm_prompt_artifact, function, graph, model_runner_step = (
-            setup_remote_model_test(project=project, model_url=model_url)
+            self._setup_remote_model_test()
         )
         model_runner_step.add_model(
             model_class="MyOpenAILLM",
